Corretta gestione del logo assente nella generazione degli avvisi di pagamento.

+++ govpay-orm/src/main/java/it/govpay/orm/model/VersamentoIncassoModel.java
@@ -94,7 +94,7 @@ public class VersamentoIncassoModel extends AbstractModel<VersamentoIncasso> {
 		this.IMPORTO_INCASSATO = new Field("importoIncassato",java.lang.Double.class,"VersamentoIncasso",VersamentoIncasso.class);
 		this.STATO_PAGAMENTO = new Field("statoPagamento",java.lang.String.class,"VersamentoIncasso",VersamentoIncasso.class);
 		this.IUV_PAGAMENTO = new Field("iuvPagamento",java.lang.String.class,"VersamentoIncasso",VersamentoIncasso.class);
-		this.SMART_ORDER_DATE = new Field("smartOrderDate",java.lang.Integer.class,"VersamentoIncasso",VersamentoIncasso.class);
+		this.SMART_ORDER_DATE = new Field("smartOrderDate",long.class,"VersamentoIncasso",VersamentoIncasso.class);
 		this.SMART_ORDER_RANK = new Field("smartOrderRank",java.lang.Integer.class,"VersamentoIncasso",VersamentoIncasso.class);
 	
 	}
@@ -156,7 +156,7 @@ public class VersamentoIncassoModel extends AbstractModel<VersamentoIncasso> {
 		this.IMPORTO_INCASSATO = new ComplexField(father,"importoIncassato",java.lang.Double.class,"VersamentoIncasso",VersamentoIncasso.class);
 		this.STATO_PAGAMENTO = new ComplexField(father,"statoPagamento",java.lang.String.class,"VersamentoIncasso",VersamentoIncasso.class);
 		this.IUV_PAGAMENTO = new ComplexField(father,"iuvPagamento",java.lang.String.class,"VersamentoIncasso",VersamentoIncasso.class);
-		this.SMART_ORDER_DATE = new ComplexField(father,"smartOrderDate",java.lang.Integer.class,"VersamentoIncasso",VersamentoIncasso.class);
+		this.SMART_ORDER_DATE = new ComplexField(father,"smartOrderDate",long.class,"VersamentoIncasso",VersamentoIncasso.class);
 		this.SMART_ORDER_RANK = new ComplexField(father,"smartOrderRank",java.lang.Integer.class,"VersamentoIncasso",VersamentoIncasso.class);
 	
 	}
